Add files via upload
Install XAMPP Server or WAMP server to run php files and install MySQL Server. Index.html HTML file allows a user to enter a task name and date for the task to be executed. The task is handled by script.js JavaScript file which allows the tasks to be displayed in a list and save the task in MySQL by posting through save_task.php PHP file. View_tasks.php PHP file allows a user to view saved tasks and beeps to alert the user that task's time has passed when a user moves cursor over the task. style.css CSS file enhances the display of the index.html file